Refactored all services, controllers for correct auth, error handling, code style and optimisations. Added error handling middleware

+++ backend/Pawfect-Pet-Adoption-App-API/Services/AnimalServices/AnimalService.cs
@@ -1,8 +1,12 @@
-﻿using AutoMapper;
+﻿using Amazon.S3.Transfer;
+using AutoMapper;
 
 using Pawfect_Pet_Adoption_App_API.Builders;
+using Pawfect_Pet_Adoption_App_API.Censors;
 using Pawfect_Pet_Adoption_App_API.Data.Entities;
 using Pawfect_Pet_Adoption_App_API.Data.Entities.EnumTypes;
+using Pawfect_Pet_Adoption_App_API.Data.Entities.Types.Authorisation;
+using Pawfect_Pet_Adoption_App_API.Exceptions;
 using Pawfect_Pet_Adoption_App_API.Models.Animal;
 using Pawfect_Pet_Adoption_App_API.Models.File;
 using Pawfect_Pet_Adoption_App_API.Models.Lookups;
@@ -10,6 +14,8 @@ using Pawfect_Pet_Adoption_App_API.Query;
 using Pawfect_Pet_Adoption_App_API.Query.Queries;
 using Pawfect_Pet_Adoption_App_API.Repositories.Implementations;
 using Pawfect_Pet_Adoption_App_API.Repositories.Interfaces;
+using Pawfect_Pet_Adoption_App_API.Services.AuthenticationServices;
+using Pawfect_Pet_Adoption_App_API.Services.BreedServices;
 using Pawfect_Pet_Adoption_App_API.Services.Convention;
 using Pawfect_Pet_Adoption_App_API.Services.FileServices;
 
@@ -21,9 +27,13 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 		private readonly IAnimalRepository _animalRepository;
 		private readonly IMapper _mapper;
 		private readonly IConventionService _conventionService;
-		private readonly Lazy<IFileService> _fileService;
+        private readonly IAuthorisationContentResolver _authorisationContentResolver;
+        private readonly Lazy<IFileService> _fileService;
+        private readonly ICensorFactory _censorFactory;
+        private readonly AuthContextBuilder _contextBuilder;
         private readonly IQueryFactory _queryFactory;
         private readonly IBuilderFactory _builderFactory;
+        private readonly IAuthorisationService _authorisationService;
 
         public AnimalService
 			(
@@ -31,9 +41,13 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 				IAnimalRepository animalRepository,
 				IMapper mapper,
 				Lazy<IFileService> fileService,
-				IQueryFactory queryFactory,
+				ICensorFactory censorFactory,
+                AuthContextBuilder contextBuilder,
+                IQueryFactory queryFactory,
                 IBuilderFactory builderFactory,
-                IConventionService conventionService
+                IAuthorisationService authorisationService,
+                IConventionService conventionService,
+				IAuthorisationContentResolver authorisationContentResolver
 
             )
 		{
@@ -41,23 +55,28 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 			_animalRepository = animalRepository;
 			_mapper = mapper;
 			_conventionService = conventionService;
-			_fileService = fileService;
+            _authorisationContentResolver = authorisationContentResolver;
+            _fileService = fileService;
+            _censorFactory = censorFactory;
+            _contextBuilder = contextBuilder;
+            _authorisationService = authorisationService;
             _queryFactory = queryFactory;
             _builderFactory = builderFactory;
 		}
-		public async Task<AnimalDto?> Persist(AnimalPersist persist, List<String> fields)
+		public async Task<Models.Animal.Animal?> Persist(AnimalPersist persist, List<String> fields)
 		{
 			Boolean isUpdate = _conventionService.IsValidId(persist.Id);
-			Animal data = new Animal();
+            Data.Entities.Animal data = new Data.Entities.Animal();
 			String dataId = String.Empty;
 
-			//*TODO* Add authorization service with user roles and permissions
+			if (!await this.AuthoriseAnimalPersist(persist, Permission.EditAnimals))
+                throw new ForbiddenException("You are not authorized to edit animals", typeof(Data.Entities.Animal), Permission.EditAnimals);
 
-			if (isUpdate)
+            if (isUpdate)
 			{
 				data = await _animalRepository.FindAsync(x => x.Id == persist.Id);
 
-				if (data == null) throw new InvalidDataException("No entity found with id given");
+				if (data == null) throw new NotFoundException("No animal found to persist", persist.Id, typeof(Data.Entities.Animal));
 
 				_mapper.Map(persist, data);
 				data.UpdatedAt = DateTime.UtcNow;
@@ -76,9 +95,7 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 			else dataId = await _animalRepository.AddAsync(data);
 
 			if (String.IsNullOrEmpty(dataId))
-			{
-				throw new InvalidOperationException("Αποτυχία κατα το Persisting");
-			}
+				throw new InvalidOperationException("Failed to persist animal");
 
 			// Return dto model
 			AnimalLookup lookup = new AnimalLookup();
@@ -87,17 +104,32 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 			lookup.Offset = 0;
 			lookup.PageSize = 1;
 
-			return (await _builderFactory.Builder<AnimalBuilder>().BuildDto(await lookup.EnrichLookup(_queryFactory).CollectAsync(), fields)).FirstOrDefault();
+            AuthContext context = _contextBuilder.OwnedFrom(lookup).AffiliatedWith(lookup).Build();
+            List<String> censoredFields = await _censorFactory.Censor<AnimalCensor>().Censor([.. lookup.Fields], context);
+            if (censoredFields.Count == 0) throw new ForbiddenException("Unauthorised access when querying animals");
+
+            lookup.Fields = censoredFields;
+            return (await _builderFactory.Builder<AnimalBuilder>().Authorise(AuthorizationFlags.OwnerOrPermissionOrAffiliation)
+										.Build(await lookup.EnrichLookup(_queryFactory).Authorise(AuthorizationFlags.OwnerOrPermissionOrAffiliation).CollectAsync(), censoredFields))
+										.FirstOrDefault();
 		}
 
-		private async Task PersistFiles(List<String> attachedFilesIds, List<String> currentFileIds)
+		private async Task<Boolean> AuthoriseAnimalPersist(AnimalPersist animal, String permission)
+		{
+			String userShelterId = await _authorisationContentResolver.CurrentPrincipalShelter();
+			if (!_conventionService.IsValidId(userShelterId) || !animal.ShelterId.Equals(userShelterId)) return false;
+
+            return await _authorisationService.AuthorizeAsync(permission);
+        }
+
+        private async Task PersistFiles(List<String> attachedFilesIds, List<String> currentFileIds)
 		{
 			// Make nul lto an empty list so that we can delete all current file Ids
 			if (attachedFilesIds == null) { attachedFilesIds = new List<String>(); }
 
-			if (currentFileIds != null)
+            if (currentFileIds != null)
 			{
-				List<String> diff = currentFileIds.Except(attachedFilesIds).ToList();
+				List<String> diff = [..currentFileIds.Except(attachedFilesIds)];
 
 				// If no difference with current , return
 				if (diff.Count == 0 && currentFileIds.Count == attachedFilesIds.Count) { return; }
@@ -116,7 +148,7 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 			lookup.PageSize = attachedFilesIds.Count;
 
 			List<Data.Entities.File> attachedFiles = await lookup.EnrichLookup(_queryFactory).CollectAsync();
-			if (attachedFiles == null || !attachedFiles.Any())
+			if (attachedFiles == null || attachedFiles.Count == 0)
 			{
 				_logger.LogError("Failed to saved attached files. No return from query");
 				return;
@@ -132,7 +164,7 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 			await _fileService.Value.Persist
 			(
 				persistModels,
-				new List<String>() { nameof(FileDto.Id) }
+				new List<String>() { nameof(Models.File.File.Id) }
 			);
 		}
 
@@ -140,18 +172,41 @@ namespace Pawfect_Pet_Adoption_App_API.Services.AnimalServices
 
 		public async Task Delete(List<String> ids)
 		{
-			// TODO : Authorization
+			if (!await this.AuthoriseAnimalDeletion(ids))
+                throw new ForbiddenException("You are not authorized to delete animals", typeof(Data.Entities.Animal), Permission.DeleteAnimals);
 
-			FileLookup lookup = new FileLookup();
-			lookup.OwnerIds = ids;
-			lookup.Fields = new List<String> { nameof(AnimalDto.Id) };
-			lookup.Offset = 0;
-			lookup.PageSize = 50;
+            FileLookup fLookup = new FileLookup();
+            fLookup.OwnerIds = ids;
+            fLookup.Fields = new List<String> { nameof(Models.File.File.Id) };
+            fLookup.Offset = 0;
+            fLookup.PageSize = 10000;
 
-			List<Data.Entities.File> attachedFiles = await lookup.EnrichLookup(_queryFactory).CollectAsync();
-			await _fileService.Value.Delete(attachedFiles?.Select(x => x.Id).ToList());
+			List<Data.Entities.File> attachedFiles = await fLookup.EnrichLookup(_queryFactory).CollectAsync();
+			await _fileService.Value.Delete([..attachedFiles?.Select(x => x.Id)]);
 
 			await _animalRepository.DeleteAsync(ids);
 		}
+
+		private async Task<Boolean> AuthoriseAnimalDeletion(List<String> animalIds)
+		{
+			if (await _authorisationService.AuthorizeAsync(Permission.DeleteAnimals)) return true;
+
+            String shelterId = await _authorisationContentResolver.CurrentPrincipalShelter();
+			if (!_conventionService.IsValidId(shelterId)) return false;
+
+            AnimalLookup lookup = new AnimalLookup();
+            lookup.Ids = animalIds;
+            lookup.ShelterIds = [shelterId];
+            lookup.Fields = new List<String> { nameof(Models.Animal.Animal.Id) };
+            lookup.Offset = 1;
+            lookup.PageSize = 10000;
+
+            List<String> allowedAnimalsToDelete = [.. (await lookup.EnrichLookup(_queryFactory).CollectAsync())?.Select(animal => animal.Id)];
+			if (allowedAnimalsToDelete == null) return false;
+
+            List<String> filteredAnimalDeletions = [.. animalIds.Except(allowedAnimalsToDelete)];
+
+			return filteredAnimalDeletions.Count == 0;
+        } 
 	}
 }
