+++ backend/Pawfect-Pet-Adoption-App-API/Services/BreedServices/BreedService.cs
@@ -1,14 +1,13 @@
 ﻿using AutoMapper;
-using Microsoft.AspNetCore.SignalR;
 using Pawfect_Pet_Adoption_App_API.Builders;
-using Pawfect_Pet_Adoption_App_API.Data.Entities;
-using Pawfect_Pet_Adoption_App_API.Models.AdoptionApplication;
+using Pawfect_Pet_Adoption_App_API.Censors;
+using Pawfect_Pet_Adoption_App_API.Data.Entities.Types.Authorisation;
+using Pawfect_Pet_Adoption_App_API.Exceptions;
 using Pawfect_Pet_Adoption_App_API.Models.Breed;
 using Pawfect_Pet_Adoption_App_API.Models.Lookups;
 using Pawfect_Pet_Adoption_App_API.Query;
-using Pawfect_Pet_Adoption_App_API.Query.Queries;
-using Pawfect_Pet_Adoption_App_API.Repositories.Implementations;
 using Pawfect_Pet_Adoption_App_API.Repositories.Interfaces;
+using Pawfect_Pet_Adoption_App_API.Services.AuthenticationServices;
 using Pawfect_Pet_Adoption_App_API.Services.Convention;
 
 namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
@@ -20,6 +19,9 @@ namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
 		private readonly IConventionService _conventionService;
         private readonly IQueryFactory _queryFactory;
         private readonly IBuilderFactory _builderFactory;
+        private readonly IAuthorisationService _authorisationService;
+        private readonly AuthContextBuilder _contextBuilder;
+        private readonly ICensorFactory _censorFactory;
 
         public BreedService
 		(
@@ -27,7 +29,10 @@ namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
 			IMapper mapper,
 			IConventionService conventionService,
 			IQueryFactory queryFactory,
-            IBuilderFactory builderFactory
+            IBuilderFactory builderFactory,
+			IAuthorisationService authorisationService,
+			AuthContextBuilder contextBuilder,
+			ICensorFactory censorFactory
         )
 		{
 			_breedRepository = breedRepository;
@@ -35,18 +40,24 @@ namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
 			_conventionService = conventionService;
             _queryFactory = queryFactory;
             _builderFactory = builderFactory;
+            _authorisationService = authorisationService;
+            _contextBuilder = contextBuilder;
+            _censorFactory = censorFactory;
         }
 
-		public async Task<BreedDto?> Persist(BreedPersist persist, List<String> fields)
+		public async Task<Models.Breed.Breed?> Persist(BreedPersist persist, List<String> fields)
 		{
-			Boolean isUpdate = _conventionService.IsValidId(persist.Id);
-			Breed data = new Breed();
+			if (!await _authorisationService.AuthorizeAsync(Permission.EditBreeds))
+				throw new ForbiddenException("You are not authorized to edit breeds", typeof(Data.Entities.Breed), Permission.EditBreeds);
+
+            Boolean isUpdate = _conventionService.IsValidId(persist.Id);
+            Data.Entities.Breed data = new Data.Entities.Breed();
 			String dataId = String.Empty;
 			if (isUpdate)
 			{
 				data = await _breedRepository.FindAsync(x => x.Id == persist.Id);
 
-				if (data == null) throw new InvalidDataException("No entity found with id given");
+				if (data == null) throw new NotFoundException("No breed found with this id to persist", persist.Id, typeof(Data.Entities.Breed));
 
 				_mapper.Map(persist, data);
 				data.UpdatedAt = DateTime.UtcNow;
@@ -63,9 +74,7 @@ namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
 			else dataId = await _breedRepository.AddAsync(data);
 
 			if (String.IsNullOrEmpty(dataId))
-			{
-				throw new InvalidOperationException("Αποτυχία κατά το persist της φυλής");
-			}
+				throw new InvalidOperationException("Failed to persist breed");
 
 			// Return dto model
 			BreedLookup lookup = new BreedLookup();
@@ -74,15 +83,24 @@ namespace Pawfect_Pet_Adoption_App_API.Services.BreedServices
 			lookup.Offset = 0;
 			lookup.PageSize = 1;
 
-			return (await _builderFactory.Builder<BreedBuilder>().BuildDto(await lookup.EnrichLookup(_queryFactory).CollectAsync(), fields)).FirstOrDefault();
+            AuthContext context = _contextBuilder.OwnedFrom(lookup).AffiliatedWith(lookup).Build();
+            List<String> censoredFields = await _censorFactory.Censor<BreedCensor>().Censor([.. lookup.Fields], context);
+            if (censoredFields.Count == 0) throw new ForbiddenException("Unauthorised access when querying breeds");
+
+            lookup.Fields = censoredFields;
+            return (await _builderFactory.Builder<BreedBuilder>().Authorise(AuthorizationFlags.OwnerOrPermissionOrAffiliation)
+					.Build(await lookup.EnrichLookup(_queryFactory).Authorise(AuthorizationFlags.OwnerOrPermissionOrAffiliation).CollectAsync(), censoredFields))
+					.FirstOrDefault();
         }
 
-		public async Task Delete(String id) { await this.Delete(new List<String>() { id }); }
+		public async Task Delete(String id) => await this.Delete(new List<String>() { id });
 
 		public async Task Delete(List<String> ids)
 		{
-			// TODO : Authorization
-			await _breedRepository.DeleteAsync(ids);
+            if (!await _authorisationService.AuthorizeAsync(Permission.DeleteBreeds))
+                throw new ForbiddenException("You are not authorized to delete breeds", typeof(Data.Entities.Breed), Permission.DeleteBreeds);
+
+            await _breedRepository.DeleteAsync(ids);
 		}
 	}
 }
